更改attach db sp命民規則 修改 遊戲邏輯結構 新增 輪盤初始化

+++ game/gameinfo.go
@@ -6,7 +6,7 @@ import (
 )
 
 type gameInfo struct {
-	FreeCount int64 `json:"freecount"`
+	FreeCount int `json:"freecount"`
 }
 
 // ToJSONStr ...
